refactor(core): modularize linter architecture
