+++ tests/core/rule-registry.test.ts
@@ -1,0 +1,17 @@
+import test from 'node:test';
+import assert from 'node:assert/strict';
+import { RuleRegistry } from '../../src/core/rule-registry.ts';
+import type { Config } from '../../src/core/linter.ts';
+
+test('RuleRegistry enables configured rules', async () => {
+  const config: Config = {
+    tokens: {},
+    rules: { 'design-token/colors': 'warn' },
+  };
+  const registry = new RuleRegistry(config);
+  await registry.load();
+  const enabled = registry.getEnabledRules();
+  assert.equal(enabled.length, 1);
+  assert.equal(enabled[0].rule.name, 'design-token/colors');
+  assert.equal(enabled[0].severity, 'warn');
+});
